healthcare-frontend/requirements.txt
streamlit>=1.30.0
streamlit-authenticator>=0.2.3
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
bcrypt>=4.1.2
